# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "VertexBPMN.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/camunda/process-definition/{id}": {
      "get": {
        "tags": [
          "CamundaProcessDefinition"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDefinitionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDefinitionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDefinitionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/camunda/process-definition/{id}/xml": {
      "get": {
        "tags": [
          "CamundaProcessDefinition"
        ],
        "summary": "Returns the BPMN XML for a process definition (bpmn-js compatible).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "CamundaProcessDefinition"
        ],
        "summary": "Updates the BPMN XML for a process definition (bpmn-js save).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateXmlRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateXmlRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateXmlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/camunda/task/{id}": {
      "get": {
        "tags": [
          "CamundaTask"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/camunda/task/{id}/form-schema": {
      "get": {
        "tags": [
          "CamundaTask"
        ],
        "summary": "Returns the form schema for a user task (form-js compatible).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "CamundaTask"
        ],
        "summary": "Updates the form schema for a user task (form-js save).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFormSchemaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFormSchemaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFormSchemaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Decision/deploy": {
      "post": {
        "tags": [
          "Decision"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeployRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Decision/by-key": {
      "get": {
        "tags": [
          "Decision"
        ],
        "parameters": [
          {
            "name": "decisionKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDefinition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDefinition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDefinition"
                }
              }
            }
          }
        }
      }
    },
    "/api/Decision/evaluate": {
      "post": {
        "tags": [
          "Decision"
        ],
        "summary": "Evaluates a DMN decision by key with input variables.",
        "description": "Example request:\r\n            \r\n    POST /api/decision/evaluate\r\n    {\r\n        \"DecisionKey\": \"my-decision\",\r\n        \"Inputs\": { \"input1\": \"value\" }\r\n    }",
        "requestBody": {
          "description": "Evaluation request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Repository": {
      "get": {
        "tags": [
          "Repository"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessDefinition"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessDefinition"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessDefinition"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Repository"
        ],
        "summary": "Deploys a new BPMN process definition.",
        "description": "Example request:\r\n            \r\n    POST /api/repository\r\n    {\r\n        \"bpmnXml\": \"<definitions ...>...</definitions>\",\r\n        \"name\": \"hello-world.bpmn\",\r\n        \"tenantId\": null\r\n    }",
        "requestBody": {
          "description": "Deployment request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryDeployRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryDeployRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryDeployRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDefinition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDefinition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDefinition"
                }
              }
            }
          }
        }
      }
    },
    "/api/Repository/{id}": {
      "get": {
        "tags": [
          "Repository"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDefinition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDefinition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDefinition"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Repository"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Task/{id}": {
      "get": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionDefinition": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "dmnXml": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DecisionResult": {
        "type": "object",
        "properties": {
          "outputs": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DeployRequest": {
        "type": "object",
        "properties": {
          "decisionKey": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "dmnXml": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EvaluateRequest": {
        "type": "object",
        "properties": {
          "decisionKey": {
            "type": "string",
            "nullable": true
          },
          "inputs": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProcessDefinition": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "bpmnXml": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ProcessDefinitionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "key": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "resource": {
            "type": "string",
            "nullable": true
          },
          "deploymentId": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "suspended": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "RepositoryDeployRequest": {
        "type": "object",
        "properties": {
          "bpmnXml": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "processInstanceId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "assignee": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "formKey": {
            "type": "string",
            "nullable": true
          },
          "formSchema": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "assignee": {
            "type": "string",
            "nullable": true
          },
          "owner": {
            "type": "string",
            "nullable": true
          },
          "taskDefinitionKey": {
            "type": "string",
            "nullable": true
          },
          "processInstanceId": {
            "type": "string",
            "nullable": true
          },
          "processDefinitionId": {
            "type": "string",
            "nullable": true
          },
          "executionId": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "due": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "followUp": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "variables": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          },
          "formKey": {
            "type": "string",
            "nullable": true
          },
          "formSchema": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateFormSchemaRequest": {
        "type": "object",
        "properties": {
          "formKey": {
            "type": "string",
            "nullable": true
          },
          "schema": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateXmlRequest": {
        "type": "object",
        "properties": {
          "bpmnXml": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}